Restructure public/editor module
