+++ public/editor/scripts/filesystem-sync/state.js
