+++ public/editor/scripts/lib/keyhandler.js
