+++ public/editor/scripts/lib/logger.js
