+++ public/editor/scripts/lib/startup.js
